requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.4
Deprecated==1.2.13
Flask==2.0.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.30
pycrypto==2.6.1
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.14.1
zipp==3.8.1
